# tsdb: use smaller allocation in blockBaseSeriesSet